Split gated builtin type names
